// apps/webapp/src/app/core/api/models/role-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PermissionDto } from '../models/permission-dto';
export interface RoleDto {

  /**
   * Creation date.
   */
  created_at: string;

  /**
   * The description of the role.
   */
  description?: string;

  /**
   * The name of the role.
   */
  name: string;

  /**
   * List of permissions assigned to this role.
   */
  permissions?: Array<PermissionDto>;

  /**
   * The unique identifier of the role.
   */
  role_id: string;

  /**
   * The tenant ID this role belongs to.
   */
  tenant_id: string;
}